Move reco algorithms into their own folder
