+++ mangaki/mangaki/utils/algos/__init__.py
@@ -1,0 +1,8 @@
+from .als import MangakiALS
+from .wals import MangakiWALS
+from .svd import MangakiSVD
+from .knn import MangakiKNN
+from .efa import MangakiEFA
+from .pca import MangakiPCA
+from .nmf import MangakiNMF
+from .zero import MangakiZero
